UI for edit journal done

+++ app/src/main/java/com/trustathanas/journalapp/Rooms/JournalEntity.kt
@@ -7,7 +7,7 @@ import android.arch.persistence.room.PrimaryKey
 @Entity(tableName = "journal_entries")
 data class JournalEntity(
         @PrimaryKey(autoGenerate = true)
-        @ColumnInfo(name = "id") val id: Long? = null,
+        @ColumnInfo(name = "id") val id: Int? = null,
         @ColumnInfo(name = "title") val title: String,
         @ColumnInfo(name = "contents") val contents: String
